+++ app/assets/stylesheets/application.css
@@ -1,6 +1,3 @@
-/*
- *= require font-awesome
- */
 
 @import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;500;800&family=Poppins:wght@300;400&display=swap');
