--- app/controllers/home_controllers.rb
@@ -1,5 +1,0 @@
-class HomeController < ApplicationController
-  skip_before_action :authenticate_user!, only: :index
-
-  def index; end
-end
