+++ app/helpers/categories_helper.rb
@@ -1,0 +1,2 @@
+module CategoriesHelper
+end
