+++ app/helpers/transaction_helper.rb
@@ -1,0 +1,2 @@
+module TransactionHelper
+end
