+++ app/views/devise/sessions/new.html.erb
@@ -1,0 +1,33 @@
+<header class="header flex">
+  <span>back</span>
+  <h2>Login</h2>
+  <span>
+    <%= link_to "Sign up", new_registration_path(resource_name), class: 'header-action' %>
+  </span>
+</header>
+<section class="access flex">
+  <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
+    <div class="field">
+      <%= f.email_field :email, autofocus: true, autocomplete: "email", placeholder: "E mail" %>
+    </div>
+
+    <div class="field">
+      <%= f.password_field :password, autocomplete: "current-password", placeholder: "Password" %>
+    </div>
+
+    <% if devise_mapping.rememberable? %>
+      <div class="field">
+        <%= f.check_box :remember_me %>
+        <%= f.label :remember_me %>
+      </div>
+    <% end %>
+
+    <div class="actions">
+      <%= f.submit "Log in", class: "submit" %>
+    </div>
+  <% end %>
+  <div class="options flex">
+    <%= link_to "Forgot your password?", new_password_path(resource_name) %>
+    <%= link_to "Didn't receive confirmation instructions?", new_category_path(resource_name) %>
+  </div>
+</section>
