+++ config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
+  
   devise_for :users, sign_out_via: [:get]
 
   root to: 'home#index'
